// backup/src/types.ts
export type ORDER = {
  HORIZONTAL: 'horizontal'
  VERTICAL: 'vertical'
  GRID: 'grid'
}

export type Point = { x: number, y: number }

export type FocusableWithKind = {
  id: string | null | undefined
  kind: 'group' | 'element'
}

export type FocusableByDirection = {
  [key in Direction]?: string | FocusableWithKind | null
}

export type ElementByDirection = {
  [key in Direction]?: string | null
}

type EventResult<T> = {
  current: T | undefined | null
  direction: Direction | undefined | null
}

export type FocusEventResult<T> = EventResult<T> & {
  prev: T | undefined | null
}

export type BlurEventResult<T> = EventResult<T> & {
  next: T | undefined | null
}

export type Focusable = {
  id: string
  /**
   * @deprecated
   * For test purposes only.
   */
  _ref?: HTMLElement
}

export type FocusableElement = Focusable & {
  instance?: TextInput | TouchableHighlight | TouchableOpacity
  group: string
  /**
   * If group is setted byOrder, the order will be used to find the next element.
   * nextByDirection will be overrided by this property.
   */
  order?: number
  nextByDirection?: FocusableByDirection
  onFocus?: (result: FocusEventResult<FocusableElement>) => void
  onBlur?: (result: BlurEventResult<FocusableElement>) => void
}

export type FocusableElementOptions = Omit<FocusableElement, '_ref' | 'group' | 'id'>

export type FocusableGroup = Focusable & {
  elements: Set<string>
}

export type FocusableGroupConfig = Focusable & {
  instance?: View
  firstElement?: string
  lastElement?: string
  nextGroupByDirection?: ElementByDirection
  byOrder?: ORDER[keyof ORDER]
  cols?: number | Record<number, number>
  saveLast?: boolean
  viewportSafe?: boolean
  threshold?: number
  onFocus?: (result: FocusEventResult<FocusableGroupConfig>) => void
  onBlur?: (result: BlurEventResult<FocusableGroupConfig>) => void
  keepFocus?: boolean
  arrowDebounce?: boolean
}

export type FocusableGroupOptions = Omit<FocusableGroupConfig, '_ref' | 'id'>

export type Rect = {
  x: number
  y: number
  width: number
  height: number
  left: number
  top: number
  right: number
  bottom: number
}

export type FocusNodeOptions = {
  preventScroll?: boolean
}

export type Adapter = {
  type: 'web' | 'react-native'
  getNodeRect: (focusable: FocusableElement | FocusableGroupConfig) => Rect
  isNodeDisabled: (focusable: FocusableElement) => boolean
  focusNode: (focusable: FocusableElement, opts?: FocusNodeOptions) => void
  isNodeFocusable: (focusable: FocusableElement) => boolean
  getNodeRef: (focusable: Focusable) => unknown,
  getFocusedNode: () => unknown
}

export type ArrowNavigationState = {
  currentElement: string | null,
  groupsConfig: Map<string, FocusableGroupConfig>
  groups: Map<string, FocusableGroup>
  elements: Map<string, FocusableElement>
  debug?: boolean
  readonly adapter: Adapter
  initialFocusElement?: string
  readonly registerCooldown?: number
}